clients/python/pyproject.toml
[project]
name = "litellm-o3-deep-research-python"
version = "0.1.0"
description = "Minimal Python example client for calling LiteLLM o3-deep-research via REST"
requires-python = ">=3.10"
dependencies = [
    "certifi>=2026.0.0",
    "python-dotenv>=1,<2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8,<9",
]

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"
